+++ ci/checks/prohibit-trailing-whitespace.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-if git grep -EIn ' +$' -- :!'*.py'; then
+if git grep -EIn ' +$' -- :!git_machete/generated_docs.py; then
   echo 'The above lines contain trailing whitespace, please tidy up'
   exit 1
 fi