+++ git_machete/generated_docs.py
@@ -249,6 +249,12 @@ long_docs: Dict[str, str] = {
               Setting this config key to `true` will force `git machete github create-pr` to take PR description
               from the message body of the first unique commit of the branch, even if `.git/info/description` and/or `.github/pull_request_template.md` is present.
 
+           `machete.github.prDescriptionIntroStyle`:
+              Select the style of the intro prepended to PR description:
+              * `full`    — include both a chain of upstream PRs (typically leading to `main`, `master`, `develop` etc.) and a tree of downstream PRs
+              * `up-only` — default, include only a chain of upstream PRs
+              * `none`    — prepend no intro to the PR description at all
+
            `machete.gitlab.{domain,remote,namespace,project}`:
 
                  GitLab self-managed domain
@@ -275,6 +281,12 @@ long_docs: Dict[str, str] = {
            `machete.gitlab.forceDescriptionFromCommitMessage`:
               Setting this config key to `true` will force `git machete gitlab create-mr` to take MR description
               from the message body of the first unique commit of the branch, even if `.git/info/description` and/or `.gitlab/merge_request_templates/Default.md` is present.
+
+           `machete.gitlab.mrDescriptionIntroStyle`:
+              Select the style of the intro prepended to MR description:
+              * `full`    — include both a chain of upstream MRs (typically leading to `main`, `master`, `develop` etc.) and a tree of downstream MRs
+              * `up-only` — default, include only a chain of upstream MRs
+              * `none`    — prepend no intro to the MR description at all
 
            `machete.overrideForkPoint.<branch>.to`:
  
@@ -656,10 +668,10 @@ long_docs: Dict[str, str] = {
  
               Sets the base of the current (or specified) branch's PR to upstream (parent) branch, as seen by git machete (see `git machete show up`).
 
-              If after changing the base the PR ends up stacked atop another PR, the PR description posted to GitHub will be prepended with a section
-              listing the entire related chain of PRs.
+              If after changing the base the PR ends up stacked atop another PR, the PR description posted to GitHub will be prepended
+              with an intro section listing the entire related chain of PRs.
 
-              This header will be updated or removed accordingly with the subsequent runs of `retarget-pr`.
+              This header will be updated or removed accordingly with the subsequent runs of `retarget-pr`, even if the base branch is already up to date.
 
               <b>Options:</b>
 
@@ -714,6 +726,12 @@ long_docs: Dict[str, str] = {
            `machete.github.forceDescriptionFromCommitMessage` (`create-pr` only):
               Setting this config key to `true` will force `git machete github create-pr` to take PR description
               from the message body of the first unique commit of the branch, even if `.git/info/description` and/or `.github/pull_request_template.md` is present.
+
+           `machete.github.prDescriptionIntroStyle` (`create-pr`, `restack-pr` and `retarget-pr`):
+              Select the style of the intro prepended to PR description:
+              * `full`    — include both a chain of upstream PRs (typically leading to `main`, `master`, `develop` etc.) and a tree of downstream PRs
+              * `up-only` — default, include only a chain of upstream PRs
+              * `none`    — prepend no intro to the PR description at all
 
         <b>Environment variables (all subcommands):</b>
            `GITHUB_TOKEN`
@@ -837,10 +855,10 @@ long_docs: Dict[str, str] = {
  
               Sets the target of the current (or specified) branch's MR to upstream (parent) branch, as seen by git machete (see `git machete show up`).
 
-              If after changing the target the MR ends up stacked atop another MR, the MR description posted to GitLab will be prepended with a section
-              listing the entire related chain of MRs.
+              If after changing the target the MR ends up stacked atop another MR, the MR description posted to GitLab will be prepended
+              with an intro section listing the entire related chain of MRs.
 
-              This header will be updated or removed accordingly with the subsequent runs of `retarget-mr`.
+              This header will be updated or removed accordingly with the subsequent runs of `retarget-mr`, even if the target branch is already up to date.
 
               <b>Options:</b>
 
@@ -882,6 +900,12 @@ long_docs: Dict[str, str] = {
            `machete.gitlab.forceDescriptionFromCommitMessage` (`create-mr` only):
               Setting this config key to `true` will force `git machete gitlab create-mr` to take MR description
               from the message body of the first unique commit of the branch, even if `.git/info/description` and/or `.gitlab/merge_request_templates/Default.md` is present.
+
+           `machete.gitlab.mrDescriptionIntroStyle` (`create-mr`, `restack-mr` and `retarget-mr`):
+              Select the style of the intro prepended to PR description:
+              * `full`    — include both a chain of upstream PRs (typically leading to `main`, `master`, `develop` etc.) and a tree of downstream PRs
+              * `up-only` — default, include only a chain of upstream PRs
+              * `none`    — prepend no intro to the PR description at all
 
         <b>Environment variables (all subcommands):</b>
            `GITLAB_TOKEN`