+++ tests/test_github_retarget_pr.py
@@ -19,7 +19,6 @@ class TestGitHubRetargetPR(BaseTest):
             mock_pr_json(head='feature_2', base='master', number=25, body=None),
             mock_pr_json(head='feature_3', base='master', number=30),
             mock_pr_json(head='feature_4', base='feature', number=35),
-            # Let's include another PR for `feature_2`, but with a different base branch
             mock_pr_json(head='feature_4', base='feature', number=40),
         )
 
@@ -119,20 +118,20 @@ class TestGitHubRetargetPR(BaseTest):
         branch_second_commit_msg = "Second commit on branch."
         (
             self.repo_sandbox.new_branch("root")
-                .commit("First commit on root.")
-                .new_branch("branch-1")
-                .commit(branch_first_commit_msg)
-                .commit(branch_second_commit_msg)
-                .push()
-                .new_branch('feature')
-                .commit('introduce feature')
-                .push()
-                .check_out('root')
-                .new_branch('branch-without-pr')
-                .commit('branch-without-pr')
-                .push()
-                .add_remote('new_origin', 'https://github.com/user/repo.git')
-                .check_out('root')
+            .commit("First commit on root.")
+            .new_branch("branch-1")
+            .commit(branch_first_commit_msg)
+            .commit(branch_second_commit_msg)
+            .push()
+            .new_branch('feature')
+            .commit('introduce feature')
+            .push()
+            .check_out('root')
+            .new_branch('branch-without-pr')
+            .commit('branch-without-pr')
+            .push()
+            .add_remote('new_origin', 'https://github.com/user/repo.git')
+            .check_out('root')
         )
 
         body: str = \
@@ -201,21 +200,21 @@ class TestGitHubRetargetPR(BaseTest):
         # branch feature present in each remote, no branch tracking data
         (
             self.repo_sandbox.remove_remote()
-                .new_branch("root")
-                .add_remote('origin_1', origin_1_remote_path)
-                .add_remote('origin_2', origin_2_remote_path)
-                .commit("First commit on root.")
-                .push(remote='origin_1')
-                .push(remote='origin_2')
-                .new_branch("branch-1")
-                .commit(branch_first_commit_msg)
-                .commit(branch_second_commit_msg)
-                .push(remote='origin_1')
-                .push(remote='origin_2')
-                .new_branch('feature')
-                .commit('introduce feature')
-                .push(remote='origin_1', set_upstream=False)
-                .push(remote='origin_2', set_upstream=False)
+            .new_branch("root")
+            .add_remote('origin_1', origin_1_remote_path)
+            .add_remote('origin_2', origin_2_remote_path)
+            .commit("First commit on root.")
+            .push(remote='origin_1')
+            .push(remote='origin_2')
+            .new_branch("branch-1")
+            .commit(branch_first_commit_msg)
+            .commit(branch_second_commit_msg)
+            .push(remote='origin_1')
+            .push(remote='origin_2')
+            .new_branch('feature')
+            .commit('introduce feature')
+            .push(remote='origin_1', set_upstream=False)
+            .push(remote='origin_2', set_upstream=False)
         )
 
         body: str = \
@@ -236,10 +235,10 @@ class TestGitHubRetargetPR(BaseTest):
         # branch feature_1 present in each remote, tracking data present
         (
             self.repo_sandbox.check_out('feature')
-                .new_branch('feature_1')
-                .commit('introduce feature 1')
-                .push(remote='origin_1')
-                .push(remote='origin_2')
+            .new_branch('feature_1')
+            .commit('introduce feature 1')
+            .push(remote='origin_1')
+            .push(remote='origin_2')
         )
 
         body = \
@@ -265,12 +264,13 @@ class TestGitHubRetargetPR(BaseTest):
 
             # Based on PR #15
 
-            ## Full chain of PRs as of 2023-12-31
+            ## Chain of upstream PRs as of 2023-12-31
 
-            * PR #20:
-              `feature_1` ➔ `feature`
             * PR #15:
-              `feature` ➔ `master`
+              `master` ← `feature`
+
+              * **PR #20 (THIS ONE)**:
+                `feature` ← `feature_1`
 
             <!-- end git-machete generated -->
             # Summary''')[1:]
@@ -278,8 +278,8 @@ class TestGitHubRetargetPR(BaseTest):
         # branch feature_2 is not present in any of the remotes
         (
             self.repo_sandbox.check_out('feature')
-                .new_branch('feature_2')
-                .commit('introduce feature 2')
+            .new_branch('feature_2')
+            .commit('introduce feature 2')
         )
 
         body = \
@@ -297,38 +297,26 @@ class TestGitHubRetargetPR(BaseTest):
         # branch feature_2 present in only one remote: origin_1 and there is no tracking data available -> infer the remote
         (
             self.repo_sandbox.check_out('feature_2')
-                .push(remote='origin_1', set_upstream=False)
+            .push(remote='origin_1', set_upstream=False)
         )
 
+        self.repo_sandbox.set_git_config_key("machete.github.prDescriptionIntroStyle", "none")
         assert_success(
             ['github', 'retarget-pr'],
             'Base branch of PR #25 has been switched to feature\n'
-            'Checking for open GitHub PRs (to determine PR chain)... OK\n'
             'Description of PR #25 has been updated\n'
         )
         pr25 = github_api_state.get_pull_by_number(25)
         assert pr25 is not None
         assert pr25['base']['ref'] == 'feature'
-        assert pr25['body'] == textwrap.dedent('''
-            <!-- start git-machete generated -->
-
-            # Based on PR #15
-
-            ## Full chain of PRs as of 2023-12-31
-
-            * PR #25:
-              `feature_2` ➔ `feature`
-            * PR #15:
-              `feature` ➔ `master`
-
-            <!-- end git-machete generated -->''')[1:]
+        assert pr25['body'] == ''
 
         # branch feature_3 present in only one remote: origin_1 and has tracking data
         (
             self.repo_sandbox.check_out('feature_2')
-                .new_branch('feature_3')
-                .commit('introduce feature 3')
-                .push(remote='origin_1')
+            .new_branch('feature_3')
+            .commit('introduce feature 3')
+            .push(remote='origin_1')
         )
 
         body = \
@@ -342,6 +330,7 @@ class TestGitHubRetargetPR(BaseTest):
             """
         rewrite_branch_layout_file(body)
 
+        self.repo_sandbox.unset_git_config_key("machete.github.prDescriptionIntroStyle")
         assert_success(
             ['github', 'retarget-pr'],
             'Base branch of PR #30 has been switched to feature_2\n'
@@ -356,14 +345,16 @@ class TestGitHubRetargetPR(BaseTest):
 
             # Based on PR #25
 
-            ## Full chain of PRs as of 2023-12-31
+            ## Chain of upstream PRs as of 2023-12-31
 
-            * PR #30:
-              `feature_3` ➔ `feature_2`
-            * PR #25:
-              `feature_2` ➔ `feature`
             * PR #15:
-              `feature` ➔ `master`
+              `master` ← `feature`
+
+              * PR #25:
+                `feature` ← `feature_2`
+
+                * **PR #30 (THIS ONE)**:
+                  `feature_2` ← `feature_3`
 
             <!-- end git-machete generated -->
             # Summary''')[1:]
@@ -393,12 +384,13 @@ class TestGitHubRetargetPR(BaseTest):
 
             # Based on PR #15
 
-            ## Full chain of PRs as of 2023-12-31
+            ## Chain of upstream PRs as of 2023-12-31
 
-            * PR #30:
-              `feature_3` ➔ `feature`
             * PR #15:
-              `feature` ➔ `master`
+              `master` ← `feature`
+
+              * **PR #30 (THIS ONE)**:
+                `feature` ← `feature_3`
 
             <!-- end git-machete generated -->
             # Summary''')[1:]
@@ -423,6 +415,39 @@ class TestGitHubRetargetPR(BaseTest):
         assert pr30 is not None
         assert pr30['base']['ref'] == 'root'
         assert pr30['body'] == '# Summary'
+
+        self.repo_sandbox.check_out('feature')
+        self.repo_sandbox.remove_remote('origin_2')
+        self.repo_sandbox.set_git_config_key("machete.github.prDescriptionIntroStyle", "full")
+
+        assert_success(
+            ['github', 'retarget-pr'],
+            'Base branch of PR #15 has been switched to branch-1\n'
+            'Checking for open GitHub PRs (to determine PR tree)... OK\n'
+            'Description of PR #15 has been updated\n'
+        )
+        pr15 = github_api_state.get_pull_by_number(15)
+        assert pr15 is not None
+        assert pr15['base']['ref'] == 'branch-1'
+        assert pr15['body'] == textwrap.dedent('''
+            <!-- start git-machete generated -->
+
+            ## Chain of upstream PRs & tree of downstream PRs as of 2023-12-31
+
+            * **PR #15 (THIS ONE)**:
+              `branch-1` ← `feature`
+
+                * PR #20:
+                  `feature` ← `feature_1`
+
+                * PR #25:
+                  `feature` ← `feature_2`
+
+                * PR #35:
+                  `feature` ← `feature_4`
+
+            <!-- end git-machete generated -->
+            # Summary''')[1:]
 
     @staticmethod
     def github_api_state_for_test_retarget_pr_root_branch() -> MockGitHubAPIState: