+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 # Release notes
 
-## New in git-machete 3.28.1
+## New in git-machete 3.29.0
+
+- added: git config keys `machete.github.prDescriptionIntroStyle` and `machete.gitlab.mrDescriptionIntroStyle`
+- added: ability to turn off PR/MR description intro completely by setting the git config keys to `none` (suggested by @tir38)
+- added: ability to also include downstream PRs/MRs in PR/MR description intro (suggested by @aouaki)
+- changed: layout and ordering of PRs/MRs in PR/MR description intro to better match `git machete status`
 
 ## New in git-machete 3.28.0